Update common and data-provider
common and data-provider migrated to Aluminium

Issue-ID: CCSDK-3111

+++ sdnr/wt/devicemanager-openroadm/provider/src/test/java/org/onap/ccsdk/features/sdnr/wt/devicemanager/openroadm/test/TestOpenRoadmNetworkElement.java
@@ -75,6 +75,7 @@ import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.yang.types.
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.data.provider.rev201110.Inventory;
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.NodeId;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 public class TestOpenRoadmNetworkElement {
 
@@ -242,7 +243,7 @@ public class TestOpenRoadmNetworkElement {
 
         Measurement measurement = mock(Measurement.class);
         PmDataType pmDataType = mock(PmDataType.class);
-        when(pmDataType.getUint64()).thenReturn(BigInteger.valueOf(2425425));
+        when(pmDataType.getUint64()).thenReturn(Uint64.valueOf(2425425));
         when(measurement.getBinNumber()).thenReturn(1452);
         when(measurement.getCompletionTime()).thenReturn(new DateAndTime("2018-10-22T15:23:43Z"));
         when(measurement.getGranularity()).thenReturn(PmGranularity._24Hour);
